Refactor : javaDoc 추가
