+++ app/build.gradle
@@ -44,6 +44,6 @@ application {
 }
 
 tasks.named('test') {
-    // Use junit platform for unit atests.
+    // Use junit platform for unit tests.
     useJUnitPlatform()
 }
